Annoying how I have to use size_t instead of int here

+++ C++/Assignment_2/main.cpp
@@ -264,7 +264,7 @@ void userOptions(User &currentUser, ReadPosts& postManager) {
             //I need to use for_each each to print out the details of each of the users here
                 //Lol, I forgot that I only need the 10 worst users
 
-            int count = min(10, postManager.getUsers().size());
+            size_t count = min(static_cast<size_t>(10), postManager.getUsers().size());
             for_each(postManager.getUsers().begin(), postManager.getUsers().begin() + count, printUser);
             break;
         case 5:
